Add functionality of + - right after x / operators
Now the calculator add correctly a + and - right after x / operators.
They also behave correctly after themselves, that is, they can be used
in the following order:
+ -> -
- -> +

However, due to the limitations of the calculator, the following would
throw a NaN:

3 + - +
The first number would be 3
The considered operator would be +
The second number is considered to be -

Since the calculator only operates two numbers at a time, the last +
would force the operation between the first and second number.
However, the second number only has sign, but not a value.
It is noted that the same would happen with the equal operation.